backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-money==1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
filelock==3.0.12
gevent==1.4.0
gevent-eventemitter==2.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pipenv==2018.11.26
pluggy==0.13.1
protobuf==3.11.3
psycopg2==2.8.4
py==1.8.1
py-moneyed==0.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
steam==0.9.1
uritemplate==3.0.1
urllib3==1.25.8
vdf==3.2
virtualenv==20.0.1
virtualenv-clone==0.5.3
wcwidth==0.1.8
whitenoise==5.0.1
